+++ spec/rubocop/cop/rspec/example_wording_spec.rb
@@ -11,7 +11,7 @@ describe RuboCop::Cop::RSpec::ExampleWording do
     expect(cop.offenses.map(&:line).sort).to eq([1])
     expect(cop.messages)
       .to eq(['Do not use should when describing your tests.'])
-    expect(cop.highlights).to eq(['it'])
+    expect(cop.highlights).to eq(['should do something'])
   end
 
   it 'skips descriptions without `should` at the beginning' do
@@ -19,5 +19,21 @@ describe RuboCop::Cop::RSpec::ExampleWording do
                          "   'here' do",
                          'end'])
     expect(cop.offenses).to be_empty
+  end
+
+  {
+    'should return something' => 'returns something',
+    'should not return something' => 'does not return something',
+    'should do nothing' => 'does nothing',
+    'should be green' => 'is green',
+    'should have sweets' => 'has sweets',
+    'should worry about the future' => 'worries about the future',
+    'should pay for pizza' => 'pays for pizza',
+    'should miss me' => 'misses me'
+  }.each do |old, new|
+    it 'autocorrects an offenses' do
+      new_source = autocorrect_source(cop, ["it '#{old}' do", 'end'])
+      expect(new_source).to eq("it '#{new}' do\nend")
+    end
   end
 end